+++ node/tests/TestUtilities.ts
@@ -305,6 +305,8 @@ export async function transactionTest(
     args.push([field + "2", field + "1"]);
     baseTransaction.sadd(key7, ["bar", "foo"]);
     args.push(2);
+    baseTransaction.sunionstore(key7, [key7, key7]);
+    args.push(2);
     baseTransaction.sinter([key7, key7]);
     args.push(new Set(["bar", "foo"]));
     baseTransaction.srem(key7, ["foo"]);